+++ src/osu/mod.rs
@@ -1,0 +1,1 @@
+pub mod osu2015;